ログインの登録処理と画面遷移を親クラスに移動させる
