+++ BookManager/LoginViewController.swift
@@ -1,7 +1,7 @@
 import UIKit
 import APIKit
 
-class LoginViewController: UIViewController {
+class LoginViewController: BaseAuthViewController {
 
     @IBOutlet weak var emailTextField: UITextField! {
         didSet {
@@ -58,17 +58,5 @@ class LoginViewController: UIViewController {
                                       message: R.string.localizable.errorLogin())
             }
         }
-    }
-
-    func saveAuthInfo(_ auth: Auth) {
-        UserDefaults.standard.set(auth.requestToken, forKey: "request_token")
-        UserDefaults.standard.set(auth.userId, forKey: "user_id")
-        UserDefaults.standard.synchronize()
-    }
-
-    func moveListViewController() {
-        let controller = R.storyboard.main.tabViewController()
-        controller?.modalTransitionStyle = .crossDissolve
-        present(controller!, animated: true, completion: nil)
     }
 }
